Reduce amount of casting

+++ src/year2015/day20.rs
@@ -3,9 +3,9 @@ const LEN: usize = 1000000;
 pub fn part1(input: &str) -> Option<usize> {
     let n: u32 = input.parse().unwrap();
     let mut vec = vec![0; LEN];
-    for i in 1..vec.len() as u32 {
-        for j in (i..vec.len() as u32).step_by(i as usize) {
-            vec[j as usize] += 10 * i;
+    for i in 1..vec.len(){
+        for j in (i..vec.len()).step_by(i) {
+            vec[j] += 10 * i as u32;
         }
     }
     vec.into_iter().position(|x| x >= n)
@@ -14,9 +14,9 @@ pub fn part1(input: &str) -> Option<usize> {
 pub fn part2(input: &str) -> Option<usize> {
     let n: u32 = input.parse().unwrap();
     let mut vec = vec![0; LEN];
-    for i in 1..vec.len() as u32 {
-        for j in (i..vec.len() as u32).step_by(i as usize).take(50) {
-            vec[j as usize] += 11 * i;
+    for i in 1..vec.len() {
+        for j in (i..vec.len()).step_by(i).take(50) {
+            vec[j] += 11 * i as u32;
         }
     }
     vec.into_iter().position(|x| x >= n)
